Put in common Xerces initialization/deinitialization and character transcoding routines for GML, NAS and GMLAS drivers
git-svn-id: https://svn.osgeo.org/gdal/trunk@35809 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/gmlas/ogrgmlaslayer.cpp
@@ -28,6 +28,7 @@
  * DEALINGS IN THE SOFTWARE.
  ****************************************************************************/
 
+// Must be first for DEBUG_BOOL case
 #include "ogr_gmlas.h"
 
 CPL_CVSID("$Id$");
@@ -301,7 +302,8 @@ void OGRGMLASLayer::PostInit( bool bIncludeGeometryXML )
             poFieldDescFeature->SetField( "field_type",
                                         oField.GetTypeName().c_str() );
         }
-        poFieldDescFeature->SetField( "field_is_list", oField.IsList() );
+        poFieldDescFeature->SetField( "field_is_list",
+                                      static_cast<int>(oField.IsList()) );
         if( oField.GetMinOccurs() != -1 )
         {
             poFieldDescFeature->SetField( "field_min_occurs",
